Create pulseaudio-microphone.sh
copied what santiprandi did here: https://github.com/polybar/polybar-scripts/tree/master/polybar-scripts/pulseaudio-microphone .
I like this script because it refreshes the mic status instantly, and for whatever reason the color foreground tags get read by polybar as if they were in the polybar config, not the script. might add a thing so that it displays the mic input sensitivity as a percentage but the important thing for me is the color switching to tell at a glance if my mic is hot or not.